Move files in agency-client, delete unused code
