+++ agency_client/src/messages/onboarding.rs
@@ -77,54 +77,3 @@ pub struct CreateAgentResponse {
     #[serde(rename = "withPairwiseDIDVerKey")]
     pub from_vk: String,
 }
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct ComMethodUpdated {
-    #[serde(rename = "@type")]
-    msg_type: MessageTypes,
-    id: String,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct UpdateComMethod {
-    #[serde(rename = "@type")]
-    msg_type: MessageTypes,
-    #[serde(rename = "comMethod")]
-    com_method: ComMethod,
-}
-
-#[derive(Debug, PartialEq)]
-pub enum ComMethodType {
-    A2A,
-    Webhook,
-}
-
-impl Serialize for ComMethodType {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error> where S: Serializer {
-        let value = match self {
-            ComMethodType::A2A => "1",
-            ComMethodType::Webhook => "2",
-        };
-        Value::String(value.to_string()).serialize(serializer)
-    }
-}
-
-impl<'de> Deserialize<'de> for ComMethodType {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error> where D: Deserializer<'de> {
-        let value = Value::deserialize(deserializer).map_err(de::Error::custom)?;
-        match value.as_str() {
-            Some("1") => Ok(ComMethodType::A2A),
-            Some("2") => Ok(ComMethodType::Webhook),
-            _ => Err(de::Error::custom("Unexpected communication method type."))
-        }
-    }
-}
-
-impl UpdateComMethod {
-    pub fn build(com_method: ComMethod) -> UpdateComMethod {
-        UpdateComMethod {
-            msg_type: MessageTypes::build(A2AMessageKinds::UpdateComMethod),
-            com_method,
-        }
-    }
-}
